exracting main+post common code
